Gestion de lieux completed...

+++ src/main/java/com/josue/kodeur/xtremanalyse/services/implementations/VilleServiceImpl.java
@@ -1,9 +1,13 @@
 package com.josue.kodeur.xtremanalyse.services.implementations;
 
+import com.josue.kodeur.xtremanalyse.dtos.QuartierDto;
 import com.josue.kodeur.xtremanalyse.entities.Prefecture;
+import com.josue.kodeur.xtremanalyse.entities.Quartier;
 import com.josue.kodeur.xtremanalyse.entities.Ville;
 import com.josue.kodeur.xtremanalyse.exceptions.NotFoundException;
+import com.josue.kodeur.xtremanalyse.mappers.MapperService;
 import com.josue.kodeur.xtremanalyse.repositories.PrefectureRepository;
+import com.josue.kodeur.xtremanalyse.repositories.QuartierRepository;
 import com.josue.kodeur.xtremanalyse.repositories.VilleRepository;
 import com.josue.kodeur.xtremanalyse.services.VilleService;
 import lombok.AllArgsConstructor;
@@ -11,7 +15,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author JosueKodeur
@@ -22,12 +28,15 @@ import java.util.List;
 @AllArgsConstructor
 public class VilleServiceImpl implements VilleService {
 
+
+    private final QuartierRepository quartierRepository;
     private final PrefectureRepository prefectureLocationRepository;
     private final VilleRepository villeLocationRepository;
+    private final MapperService mapperService;
 
     @Override
-    public Ville save(Ville ville, Long prefectureID) throws NotFoundException{
-        Prefecture prefecture = prefectureLocationRepository.findById(prefectureID)
+    public Ville save(Ville ville) throws NotFoundException{
+        Prefecture prefecture = prefectureLocationRepository.findById(ville.getPrefecture().getId())
                 .orElseThrow(() -> new NotFoundException("Prefecture Introuvable"));
         ville.setPrefecture(prefecture);
         ville.setCreatedAt(LocalDateTime.now());
@@ -40,20 +49,33 @@ public class VilleServiceImpl implements VilleService {
     public Ville update(Long ID, Ville ville) throws NotFoundException {
         Ville currentVille = villeLocationRepository.findById(ID)
                 .orElseThrow(() -> new NotFoundException("Ville Introuvable"));
+        currentVille.setPrefecture(prefectureLocationRepository.findById(ville
+                .getPrefecture()
+                .getId()).orElseThrow(() -> new NotFoundException("Prefecture introuvable")));
         currentVille.setNom(ville.getNom());
         currentVille.setNombreHabitant(ville.getNombreHabitant());
-        currentVille.setPrefecture(ville.getPrefecture());
         currentVille.setUpdateAt(LocalDateTime.now());
         return currentVille;
     }
 
     @Override
-    public void delete(Long ID) {
+    public void delete(Long ID) throws NotFoundException{
+        if (ID == null)
+            throw new NotFoundException("Ville introuvable");
         villeLocationRepository.deleteById(ID);
     }
 
     @Override
     public List<Ville> listAll() {
         return villeLocationRepository.findAll();
+    }
+
+    @Override
+    public List<QuartierDto> findQuartiersOfVille(Long ID) {
+        if (ID==null)
+            return new ArrayList<>();
+        List<Quartier> quartiers = quartierRepository.findQuartiersOfVille(ID);
+
+        return quartiers.stream().map(mapperService::toQuartierDto).collect(Collectors.toList());
     }
 }
